+++ WorldCities/WorldCitiesAPI/Data/Models/City.cs
@@ -1,0 +1,53 @@
+﻿using Microsoft.EntityFrameworkCore;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace WorldCitiesAPI.Data.Models
+{
+  [Table("Cities")]
+  [Index(nameof(Name))]
+  [Index(nameof(Lat))]
+  [Index(nameof(Lon))]
+  public class City
+  {
+    #region Properties
+    /// <summary>
+    /// The unique id and primary key for this City
+    /// </summary>
+    [Key]
+    [Required]
+    public string Id { get; set; }
+
+    /// <summary>
+    /// City name (in UTF8 format)
+    /// </summary>
+    public string Name { get; set; } = null!;
+
+    /// <summary>
+    /// City latitude
+    /// </summary>
+    [Column(TypeName = "decimal(7,4)")]
+    public double Lat {  get; set; }
+
+    /// <summary>
+    /// City longitude
+    /// </summary>
+    [Column(TypeName = "decimal(7,4)")]
+    public double Lon { get; set; }
+
+    /// <summary>
+    /// Country Id (foreign key)
+    /// </summary>
+    [ForeignKey(nameof(Country))]
+    public int CountryId { get; set; }
+    #endregion
+
+    #region Navigation Properties
+    /// <summary>
+    /// The country related to this city
+    /// </summary>
+    public Country? Country { get; set; } = null!;
+    #endregion
+  }
+
+}